# Conflicts:
#	example/package-lock.json
#	example/src/components/Pickers.js
#	example/src/components/UConfirmationButtonExample.js
#	package-lock.json
#	package.json
#	src/components/ActiveKeyboardDatePicker/ActiveKeyboardDatePicker.js
#	src/components/ActiveKeyboardTimePicker/ActiveKeyboardTimePicker.js
#	src/components/UAsyncBadge/UAsyncBadge.js
#	src/components/UConfirmationButton/UConfirmationButton.js
#	src/components/UConfirmationButton/UConfirmationButton.md
#	src/components/UDatePicker/UDatePicker.js
#	src/components/UDateTimePicker/UDateTimePicker.js
#	src/components/UKeyboardDatePicker/UKeyboardDatePicker.js
#	src/components/UKeyboardDateTimePicker/UKeyboardDateTimePicker.js
#	src/components/UKeyboardTimePicker/UKeyboardTimePicker.js
#	src/components/USearchBox/USearchBox.js
#	src/components/USelectPicker/Control.js
#	src/components/USelectPicker/Menu.js
#	src/components/USelectPicker/NoOptionsMessage.js
#	src/components/USelectPicker/USelectPicker.js
#	src/components/UTextField/UTextField.js
#	src/components/UTimePicker/UTimePicker.js
#	styleguide.config.js

+++ src/components/UConfirmationButton/UConfirmationButton.js
@@ -1,8 +1,19 @@
 import React, { useState } from 'react'
 import { styled } from '@mui/material/styles'
 import PropTypes from 'prop-types'
-import { Tooltip, IconButton, Menu, MenuItem, Box } from '@mui/material'
+import {
+  Tooltip,
+  IconButton,
+  Menu,
+  MenuItem,
+  Dialog,
+  DialogContent,
+  Typography,
+  DialogActions,
+  Box,
+} from '@mui/material'
 import DeleteOutlinedIcon from '@mui/icons-material/DeleteOutlined'
+import UButton from '../UButton'
 
 const PREFIX = 'UConfirmationButton'
 
@@ -26,6 +37,17 @@ const StyledBox = styled(Box)(({ theme }) => ({
   },
 }))
 
+const CONTROL_VARIANTS = {
+  menuItem: 'menuItem',
+  icon: 'icon',
+  button: 'button',
+}
+
+const CONFIRMATION_VARIANTS = {
+  menu: 'menu',
+  popup: 'popup',
+}
+
 /**
  * UConfirmationButton is a component to get confirmation from the user before processing it.
  * This button displays the confirmation where user has clicked the button
@@ -41,34 +63,63 @@ export default function UConfirmationButton(props) {
     icon,
     confirmText,
     confirmActionText,
-    ...others
+    cancelText,
+    buttonVariant,
+    confirmVariant,
   } = props
   const [deleteAnchorEl, setDeleteAnchorEl] = useState(null)
+  const [openDialog, setOpenDialog] = useState(false)
 
   const handleConfirm = e => {
-    onConfirm(e, id)
-    handleDeleteOptionClicked(e)
+    onConfirm && onConfirm(e, id)
+    if (confirmVariant === CONFIRMATION_VARIANTS.popup) {
+      setOpenDialog(false)
+    } else {
+      handleDeleteOptionClicked(e)
+    }
   }
 
   const handleDeleteOptionClicked = e => {
     setDeleteAnchorEl(null)
   }
+
+  const handleClick = e => {
+    if (confirmVariant === CONFIRMATION_VARIANTS.popup) {
+      setOpenDialog(true)
+    } else {
+      setDeleteAnchorEl(e.currentTarget)
+    }
+  }
+
+  const handleCancelPopup = e => {
+    setOpenDialog(false)
+  }
+
   return (
     <StyledBox>
-      {variant === 'menuItem' ? (
-        <MenuItem
-          onClick={e => setDeleteAnchorEl(e.currentTarget)}
-          className={classes.menuLabel}
-        >
+      {variant === CONTROL_VARIANTS.menuItem ? (
+        <MenuItem onClick={handleClick} className={classes.menuLabel}>
           {icon}
-          <span className={classes.span}>{buttonText}</span>
+          <Box component="span" className={icon ? classes.span : ''}>
+            {buttonText}
+          </Box>
         </MenuItem>
+      ) : variant === CONTROL_VARIANTS.button ? (
+        <UButton
+          startIcon={icon}
+          onClick={handleClick}
+          aria-controls={`delete-confirmation-menu-${id}`}
+          aria-haspopup="true"
+          variant={buttonVariant}
+        >
+          {buttonText}
+        </UButton>
       ) : (
         <Tooltip title={buttonText} placement="top">
           <IconButton
             aria-controls={`delete-confirmation-menu-${id}`}
             aria-haspopup="true"
-            onClick={e => setDeleteAnchorEl(e.currentTarget)}
+            onClick={handleClick}
             disabled={!enabled}
             className={classes.menuLabel}
             size="large"
@@ -77,18 +128,36 @@ export default function UConfirmationButton(props) {
           </IconButton>
         </Tooltip>
       )}
-      <Menu
-        id={`delete-confirmation-menu-${id}`}
-        anchorEl={deleteAnchorEl}
-        keepMounted
-        open={Boolean(deleteAnchorEl)}
-        onClose={handleDeleteOptionClicked}
-        className={classes.menuDelete}
-      >
-        <MenuItem disabled>{confirmText}</MenuItem>
-        <MenuItem onClick={e => handleConfirm(e)}>{confirmActionText}</MenuItem>
-      </Menu>
-    </StyledBox>
+      {confirmVariant === CONFIRMATION_VARIANTS.popup ? (
+        <Dialog id={`delete-confirmation-menu-${id}`} open={openDialog}>
+          <DialogContent>
+            <Typography variant="body1">{confirmText || ''}</Typography>
+          </DialogContent>
+          <DialogActions>
+            <UButton onClick={handleCancelPopup} variant="uDefault">
+              {cancelText || ''}
+            </UButton>
+            <UButton onClick={handleConfirm} variant="uPrimary" color="primary">
+              {confirmActionText || ''}
+            </UButton>
+          </DialogActions>
+        </Dialog>
+      ) : (
+        <Menu
+          id={`delete-confirmation-menu-${id}`}
+          anchorEl={deleteAnchorEl}
+          keepMounted
+          open={Boolean(deleteAnchorEl)}
+          onClose={handleDeleteOptionClicked}
+          className={classes.menuDelete}
+        >
+          <MenuItem disabled>{confirmText}</MenuItem>
+          <MenuItem onClick={e => handleConfirm(e)}>
+            {confirmActionText}
+          </MenuItem>
+        </Menu>
+      )}
+    </React.Fragment>
   )
 }
 
@@ -99,16 +168,29 @@ UConfirmationButton.propTypes = {
   buttonText: PropTypes.string,
   /** trigger confirmation function */
   onConfirm: PropTypes.func.isRequired,
-  /** variant: menu or icon */
-  variant: PropTypes.oneOf(['menuItem', 'icon']),
+  /** variant: menuItem or icon or button */
+  variant: PropTypes.oneOf([
+    CONTROL_VARIANTS.menuItem,
+    CONTROL_VARIANTS.icon,
+    CONTROL_VARIANTS.button,
+  ]),
+  /** Confirm variant: popup or menu */
+  confirmVariant: PropTypes.oneOf([
+    CONFIRMATION_VARIANTS.popup,
+    CONFIRMATION_VARIANTS.menu,
+  ]),
   /** if the variant is menuitem, this prop make sure the item enable or not */
   enabled: PropTypes.bool,
+  /** button cancel text for popup variant */
+  cancelText: PropTypes.string,
   /** button confirm text */
   confirmText: PropTypes.string,
   /** confirm action text */
   confirmActionText: PropTypes.string,
   /** custom icon */
   icon: PropTypes.element,
+  /** Button variant applied to menuItem button */
+  buttonVariant: PropTypes.string,
 }
 
 UConfirmationButton.defaultProps = {
@@ -117,5 +199,8 @@ UConfirmationButton.defaultProps = {
   enabled: true,
   confirmText: 'Confirm deletion?',
   confirmActionText: 'Yes, delete',
+  cancelText: 'No',
   icon: <DeleteOutlinedIcon />,
+  buttonVariant: 'text',
+  confirmVariant: 'menu',
 }
